Configure all remaining JPA entities to match PostgreSQL database schema
CREATED ENTITIES (8):
- Azienda: Company data with address, email, phone information
- Tirocinio (Internship): Links allievo to azienda, tracks dates and stato_esito
- Responsabile (Supervisor): Links to Utente, tracks tipo_responsabile enum
- ContattoAziendale (Company Contact): Links azienda to utente, tracks ruolo_contatto enum
- ColloquioTirocinio (Interview): Links allievo and azienda with CASCADE delete
- DocumentoTirocinio (Internship Document): Tracks tipo_documento enum for tirocinio
- Monitoraggio (Monitoring): Tracks supervision activities for tirocinio
- CasoCritico (Critical Case): Tracks issues/alerts for allievo with CASCADE delete

CREATED ENUMS (4):
- StatoEsito: IN_ATTESA, POSITIVO, NEGATIVO, RITIRATO
- TipoResponsabile: DOCENTE, TUTOR_AZIENDALE, COORDINATORE, DIRETTORE
- RuoloContatto: TITOLARE, DIRETTORE, RESPONSABILE_HR, RESPONSABILE_TIROCINI, TUTOR_AZIENDALE
- TipoDocumento: PROGETTO_FORMATIVO, FOGLIO_PRESENZE, VALUTAZIONE_FINALE, ATTESTATO, RAPPORTO_FINALE

UPDATED EXISTING ENTITIES (4):
- Allievo: Changed id LongInteger, added OneToOne to Utente, added OneToMany for Tirocinio, CasoCritico, ColloquioTirocinio
- Utente: Changed idUtente LongInteger for consistency
- Corso: Changed id LongInteger for consistency
- Ruolo: Enabled @OneToMany for Utente with @ToString.Exclude, @EqualsAndHashCode.Exclude

CREATED REPOSITORIES (8):
- AziendaRepository: findByRagioneSociale, findByPartitaIva
- TirocinioRepository: findByAllievoId, findByAziendaId, findByDateRange
- ResponsabileRepository: findByCodiceFiscale, findByEmail, findByUtenteId
- ContattoAzienddaleRepository: findByAziendaId, findByUtenteId
- ColloquioTirocinioRepository: findByAllievoId, findByAziendaId, findByDateRange
- DocumentoTirocinioRepository: findByTirocinioId
- MonitoraggioRepository: findByTirocinioId
- CasoCriticoRepository: findByAllievoId, findByRisolteFalse

All relationships properly mapped with LazyLoading and CASCADE DELETE where applicable.
Entity IDs normalized to Integer type across all entities for consistency.
Backend successfully compiles with Maven without errors.

+++ WebApplication/Backend/src/main/java/com/its/gestionale/entity/Utente.java
@@ -25,7 +25,7 @@ public class Utente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_utente")
-    private Long idUtente;
+    private Integer idUtente;
 
     @Column(unique = true, length = 100)
     private String username;
